# api/openapi.yaml
openapi: 3.1.0
info:
  title: mfd fixture lab
  version: 0.3.0
  description: Artificial experiment lab with separate protected read-only eToro Demo and Real importers. Broker order and model execution are unavailable.
servers:
  - url: https://mfd.aklein.fr
  - url: http://localhost:8088
paths:
  /api/v1/replays:
    post:
      operationId: createFixtureReplay
      summary: Enqueue a durable artificial experiment
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_-]{8,100}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Experiment'
      responses:
        '202':
          description: Queued or previously accepted under the same key. Follow status_url for the current state.
          content:
            application/json:
              schema:
                type: object
                required: [run_id, status_url]
                properties:
                  run_id: {type: string, format: uuid}
                  status_url: {type: string}
        '400': {description: Invalid key or request body}
        '403': {description: Browser cross-origin write rejected}
        '409': {description: Idempotency key already belongs to a different experiment}
        '415': {description: Content-Type must be application/json}
        '503': {description: Storage unavailable; retry with the same key}
  /api/v1/runs:
    get:
      operationId: listRuns
      summary: Latest 50 runs in descending creation order; result is null in summaries
      responses:
        '200':
          description: Run summaries
          content:
            application/json:
              schema:
                type: object
                required: [runs, limit]
                properties:
                  limit: {type: integer, const: 50}
                  runs:
                    type: array
                    items: {$ref: '#/components/schemas/Run'}
        '503': {description: Storage unavailable}
  /api/v1/runs/{id}:
    get:
      operationId: getRun
      summary: A run including captured snapshots and decisions
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: Run detail
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Run'}
        '400': {description: Malformed ID}
        '404': {description: Run not found}
        '503': {description: Storage unavailable}
  /api/v1/operations:
    get:
      operationId: inspectOperations
      summary: Read-only PostgreSQL and NATS queue view
      responses:
        '200':
          description: Dependency, storage, queue and recent job status
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Operations'}
        '503': {description: PostgreSQL unavailable}
  /api/v1/etoro/status:
    get:
      operationId: getEtoroAccountStatus
      summary: Demo and Real read status without account values
      responses:
        '200':
          description: Latest read result and encrypted snapshot count for each environment
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BrokerOverview'}
  /api/v1/etoro/snapshots:
    get:
      operationId: listEtoroSnapshots
      summary: Latest 30 decrypted portfolio summaries for one environment
      parameters:
        - {$ref: '#/components/parameters/BrokerEnvironment'}
      security: [{operatorToken: []}]
      responses:
        '200':
          description: Provider-reported values and instrument IDs for the selected environment, separate from artificial sleeves
          content:
            application/json:
              schema:
                type: object
                required: [environment, snapshots, limit]
                properties:
                  environment: {type: string, enum: [demo, real]}
                  limit: {type: integer, const: 30}
                  snapshots:
                    type: array
                    items: {$ref: '#/components/schemas/BrokerSnapshot'}
        '401': {description: Operator token required}
        '400': {description: Invalid environment}
        '503': {description: Encrypted data unavailable or invalid}
  /api/v1/etoro/sync:
    post:
      operationId: syncEtoroAccount
      summary: Request an official read-only portfolio import for one environment
      description: A denied key is checked once against the other documented endpoint and is assigned only after a valid response. At most one sync per key every 30 seconds. Denied or malformed reads create a status, never an account snapshot.
      parameters:
        - {$ref: '#/components/parameters/BrokerEnvironment'}
      security: [{operatorToken: []}]
      responses:
        '200':
          description: Connector status after attempting the read
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BrokerStatus'}
        '401': {description: Operator token required}
        '400': {description: Invalid environment}
        '503': {description: Local storage unavailable}
  /healthz:
    get:
      operationId: liveness
      responses:
        '200': {description: 'Process alive in fixture mode; includes the compiled revision'}
  /readyz:
    get:
      operationId: readiness
      responses:
        '200': {description: 'PostgreSQL, Redis and NATS available'}
        '503': {description: At least one dependency unavailable; body gives individual booleans}
  /metrics:
    get:
      operationId: metrics
      responses:
        '200': {description: Prometheus text exposition}
        '503': {description: Database metrics unavailable}
components:
  parameters:
    BrokerEnvironment:
      name: environment
      in: query
      description: Defaults to demo for older clients; specify explicitly for new clients.
      schema: {type: string, enum: [demo, real], default: demo}
  securitySchemes:
    operatorToken:
      type: apiKey
      in: header
      name: X-MFD-Operator-Token
  schemas:
    BrokerOverview:
      type: object
      required: [provider, execution_enabled, environments]
      properties:
        provider: {type: string, const: etoro}
        execution_enabled: {type: boolean, const: false}
        environments:
          type: object
          required: [demo, real]
          properties:
            demo: {$ref: '#/components/schemas/BrokerStatus'}
            real: {$ref: '#/components/schemas/BrokerStatus'}
    BrokerStatus:
      type: object
      required: [configured, provider, environment, last_attempt, last_result, http_status, last_success, snapshot_count, execution_enabled]
      properties:
        configured: {type: boolean}
        provider: {type: string, const: etoro}
        environment: {type: string, enum: [demo, real]}
        last_attempt: {type: [string, 'null'], format: date-time}
        last_result: {type: string}
        http_status: {type: integer}
        last_success: {type: [string, 'null'], format: date-time}
        snapshot_count: {type: integer}
        execution_enabled: {type: boolean, const: false}
    BrokerSnapshot:
      type: object
      required: [id, environment, fetched_at, source_sha256, portfolio]
      properties:
        id: {type: string, format: uuid}
        environment: {type: string, enum: [demo, real]}
        fetched_at: {type: string, format: date-time}
        source_sha256: {type: string, description: Hash of the encrypted source payload}
        portfolio:
          type: object
          required: [provider_at, currency, available_cash, total_value, current_pnl, instruments, mirror_count]
          properties:
            provider_at: {type: string, format: date-time}
            currency: {type: string, const: USD}
            available_cash: {$ref: '#/components/schemas/Decimal'}
            total_value: {$ref: '#/components/schemas/Decimal'}
            current_pnl: {$ref: '#/components/schemas/Decimal'}
            mirror_count: {type: integer}
            instruments:
              type: array
              items:
                type: object
                required: [instrument_id, asset_currency, net_units, exposure_usd, pnl_asset_currency, liquidation_usd, average_leverage]
                properties:
                  instrument_id: {type: integer}
                  asset_currency: {type: string}
                  net_units: {$ref: '#/components/schemas/Decimal'}
                  exposure_usd: {$ref: '#/components/schemas/Decimal'}
                  pnl_asset_currency: {$ref: '#/components/schemas/Decimal'}
                  liquidation_usd: {$ref: '#/components/schemas/Decimal'}
                  average_leverage: {$ref: '#/components/schemas/Decimal'}
    Experiment:
      type: object
      additionalProperties: false
      description: Artificial final quote shock; omitted values default to the baseline. No holdings or orders change.
      properties:
        name: {type: string, maxLength: 60, default: Baseline}
        final_shock_bps:
          type: object
          additionalProperties: false
          properties:
            AAPL: {type: integer, minimum: -5000, maximum: 5000}
            MSFT: {type: integer, minimum: -5000, maximum: 5000}
        review_threshold_bps: {type: integer, minimum: 1, maximum: 5000, default: 100}
    Operations:
      type: object
      required: [generated_at, dependencies, database, queue]
      properties:
        generated_at: {type: string, format: date-time}
        dependencies:
          type: object
          properties:
            postgres: {type: boolean}
            nats: {type: boolean}
            redis: {type: boolean}
        database:
          type: object
          properties:
            runs: {type: object, additionalProperties: {type: integer}}
            storage_bytes: {type: integer}
            outbox_pending: {type: integer}
            recent_jobs:
              type: array
              items:
                type: object
                properties:
                  run_id: {type: string, format: uuid}
                  name: {type: string}
                  status: {type: string}
                  created_at: {type: string, format: date-time}
                  published_at: {type: [string, 'null'], format: date-time}
                  finished_at: {type: [string, 'null'], format: date-time}
        queue:
          type: object
          properties:
            connected: {type: boolean}
            stream: {type: string}
            messages: {type: integer}
            bytes: {type: integer}
            consumers: {type: integer}
            pending: {type: integer}
            ack_pending: {type: integer}
            redelivered: {type: integer}
            error: {type: string}
    Decimal:
      type: string
      pattern: '^-?[0-9]+(\.[0-9]+)?$'
      description: Exact decimal serialized as a string.
    Position:
      type: object
      required: [symbol, currency, units, cost]
      properties:
        symbol: {type: string}
        currency: {type: string, const: USD}
        units: {$ref: '#/components/schemas/Decimal'}
        cost: {$ref: '#/components/schemas/Decimal'}
    Quote:
      type: object
      required: [symbol, currency, bid, ask, at]
      properties:
        symbol: {type: string}
        currency: {type: string, const: USD}
        bid: {$ref: '#/components/schemas/Decimal'}
        ask: {$ref: '#/components/schemas/Decimal'}
        at: {type: string, format: date-time}
    Snapshot:
      type: object
      required: [sleeve, at, currency, cash, market_value, equity, liquidation_equity, unrealized_pnl, positions, quotes]
      properties:
        sleeve: {type: string}
        at: {type: string, format: date-time}
        currency: {type: string, const: USD}
        cash: {$ref: '#/components/schemas/Decimal'}
        market_value: {$ref: '#/components/schemas/Decimal'}
        equity: {$ref: '#/components/schemas/Decimal'}
        liquidation_equity: {$ref: '#/components/schemas/Decimal'}
        unrealized_pnl: {$ref: '#/components/schemas/Decimal'}
        positions: {type: array, items: {$ref: '#/components/schemas/Position'}}
        quotes: {type: array, items: {$ref: '#/components/schemas/Quote'}}
    Decision:
      type: object
      required: [id, sleeve, at, actor, strategy, action, reason, checks, snapshot_index]
      properties:
        id: {type: string, description: Unique within the containing run}
        sleeve: {type: string}
        at: {type: string, format: date-time}
        actor: {type: string}
        strategy: {type: string}
        action: {type: string, enum: [hold, review]}
        reason: {type: string}
        checks: {type: array, items: {type: string}}
        snapshot_index: {type: integer, minimum: 0}
    Run:
      type: object
      required: [id, status, mode, actor, dataset_hash, experiment, created_at, finished_at, result, error]
      properties:
        id: {type: string, format: uuid}
        status: {type: string, enum: [queued, completed, failed]}
        mode: {type: string, const: fixture}
        actor: {type: string, description: Local caller label; authenticated attribution is not implemented.}
        dataset_hash: {type: string}
        experiment: {$ref: '#/components/schemas/Experiment'}
        created_at: {type: string, format: date-time}
        finished_at: {type: [string, 'null'], format: date-time}
        error: {type: [string, 'null']}
        result:
          oneOf:
            - type: 'null'
            - type: object
              required: [dataset_hash, valuation_version, experiment, snapshots, decisions]
              properties:
                dataset_hash: {type: string}
                valuation_version: {type: string}
                experiment: {$ref: '#/components/schemas/Experiment'}
                snapshots: {type: array, items: {$ref: '#/components/schemas/Snapshot'}}
                decisions: {type: array, items: {$ref: '#/components/schemas/Decision'}}